test: add graphql cooldown coverage for two-factor auth
